+++ src/CleanRegex/Match/Details/Group/CapturingGroup.php
@@ -51,10 +51,5 @@ interface CapturingGroup
     /**
      * @deprecated
      */
-    public function orElse(callable $substituteProducer);
-
-    /**
-     * @deprecated
-     */
     public function map(callable $mapper): Optional;
 }