+++ test/Fakes/CleanRegex/Match/Details/Group/TextGroup.php
@@ -107,11 +107,6 @@ class TextGroup implements Group
         throw $this->fail();
     }
 
-    public function orElse(callable $substituteProducer)
-    {
-        throw $this->fail();
-    }
-
     public function map(callable $mapper): Optional
     {
         throw $this->fail();